==> data-raw/making_of/07_MakingOfMigParlMini.Rmd <==
---
title: "MigParl 5 - Creation of a Sample Corpus"
author: "Christoph Leonhardt"
date: "03 11 2019"
output: html_document
---

## Prep

```{r setup}
library(polmineR)
corpus_new <- "MigParlMini"

s_attrs <- c("speaker", "party", "role", "lp", "session", "date", "src", "url", "regional_state", "interjection", "year", "agenda_item", "agenda_item_type", "speech", "migration_integration_probability") 

p_attrs <- c("word", "pos", "lemma", "ner")

registry <- Sys.getenv("CORPUS_REGISTRY")
if (!file.exists(registry)) stop("environment variable CORPUS_REGISTRY needs to be defined")
stateparl_registry_file <- file.path(registry, tolower(corpus_new))
if (file.exists(stateparl_registry_file)) file.remove(stateparl_registry_file)

data_dir <- file.path(dirname(registry), "indexed_corpora")
if (!file.exists(data_dir)) dir.create(data_dir)
stateparl_data_dir <- file.path(data_dir, tolower(corpus_new))
if (!file.exists(stateparl_data_dir)) dir.create(stateparl_data_dir)
file.remove(list.files(stateparl_data_dir, full.names = TRUE))
```



## Sampling

Using the entire MigParl corpus (see 03_TokenstreamToCWB.Rmd), we create a sample corpus (MigParlMini) to demonstrate some features.

```{r}
library(data.table)
library(cwbtools)
use("MigParl")
```

Let's assume that 1% of MigParl's volume would be appropriate both as an illustrative sample and in terms of file size in an R package. We want to find out the size of the new corpus should be and we want to draw a random sample from the days MigParl covers.

```{r}
size("MIGPARL") # MigParl comprises 38 million tokens
length(s_attributes("MIGPARL", "date")) # about 2100 days. 1% would be about 21 days. Let's go with 20 here.
set.seed(111)
random_days <- sample(s_attributes("MIGPARL", "date"), size = 20)
```

Next, we want to build a partition which contains the random days from above.

```{r}
date_regex <- paste0(sort(random_days), collapse = "|")
sample_partition <- partition("MIGPARL", date = date_regex, regex = TRUE) 
size(sample_partition) # 270,000 tokens
size(sample_partition)/size("MIGPARL")*100 # ~0.7%
```

## Decoding

```{r}
s_attribute_dt <- polmineR::decode(sample_partition)
s_attribute_dt[, cpos :=  as.character(as.integer(factor(cpos)) - 1)][, struc :=  as.character(as.integer(factor(struc)) - 1)]
tokenstream_dt <- copy(s_attribute_dt)
tokenstream_dt <- tokenstream_dt[, c("word", "pos", "lemma", "id", "cpos")]

s_attribute_dt[, c("word", "lemma", "pos") := NULL]
s_attribute_dt <- s_attribute_dt[,{list(cpos_left = min(as.integer(.SD[["cpos"]])), cpos_right = max(as.integer(.SD[["cpos"]])), 
                                    id = unique(.SD[["id"]]),
                                    speaker = unique(.SD[["speaker"]]),
                                    party = unique(.SD[["party"]]),
                                    role = unique(.SD[["role"]]),
                                    lp = unique(.SD[["lp"]]),
                                    src = unique(.SD[["src"]]),
                                    url = unique(.SD[["url"]]),
                                    session = unique(.SD[["session"]]),
                                    date = unique(.SD[["date"]]),
                                    interjection = unique(.SD[["interjection"]]),
                                    year = unique(.SD[["year"]]),
                                    agenda_item = unique(.SD[["agenda_item"]]),
                                    agenda_item_type = unique(.SD[["agenda_item_type"]]),
                                    speech = unique(.SD[["speech"]]),
                                    migration_integration_probability = unique(.SD[["migration_integration_probability"]]), 
                                    regional_state = unique(.SD[["regional_state"]])
  )}, by = "struc"]



```

## Corpus Encode

```{r}
CD <- CorpusData$new()
CD$tokenstream <- tokenstream_dt
CD$metadata <- s_attribute_dt
CD$metadata$lp <- as.character(CD$metadata$lp)
CD$metadata$session <- as.character(CD$metadata$session)
CD$metadata$interjection <- as.character(CD$metadata$interjection)
CD$metadata$year <- as.character(CD$metadata$year)

# fix encoding

  for (name in colnames(CD$metadata[,sapply(CD$metadata, is.character), with = FALSE])) {
    CD$metadata[[name]] <- iconv(CD$metadata[[name]], from = "latin1", to = "UTF-8")
  }


message("... start encoding")

CD$encode(
  registry_dir = registry, data_dir = stateparl_data_dir,
  corpus = toupper(corpus_new), encoding = "latin1", method = "R",
  p_attributes = p_attrs, s_attributes = s_attrs,
  compress = TRUE
)

message("... save tar.gz")

cwbtools::corpus_as_tarball(corpus=corpus_new, registry_dir = Sys.getenv('CORPUS_REGISTRY'), tarfile='~/lab/tmp/migparlmini.tar.gz')
```
